// slidetap-client/src/models/attribute_value_type.ts
//    Copyright 2024 SECTRA AB
//
//    Licensed under the Apache License, Version 2.0 (the "License");
//    you may not use this file except in compliance with the License.
//    You may obtain a copy of the License at
//
//        http://www.apache.org/licenses/LICENSE-2.0
//
//    Unless required by applicable law or agreed to in writing, software
//    distributed under the License is distributed on an "AS IS" BASIS,
//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//    See the License for the specific language governing permissions and
//    limitations under the License.

export enum AttributeValueType {
  STRING = 1,
  DATETIME = 2,
  NUMERIC = 3,
  MEASUREMENT = 4,
  CODE = 5,
  ENUM = 6,
  BOOLEAN = 7,
  OBJECT = 8,
  LIST = 10,
  UNION = 11,
}


export const AttributeValueTypeStrings = {
    [AttributeValueType.STRING]: "String",
    [AttributeValueType.DATETIME]: "Datetime",
    [AttributeValueType.NUMERIC]: "Numeric",
    [AttributeValueType.MEASUREMENT]: "Measurement",
    [AttributeValueType.CODE]: "Code",
    [AttributeValueType.ENUM]: "Enum",
    [AttributeValueType.BOOLEAN]: "Boolean",
    [AttributeValueType.OBJECT]: "Object",
    [AttributeValueType.LIST]: "List",
    [AttributeValueType.UNION]: "Union"
  }